Fixed warning messages
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@117 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/win/zfilew.c
@@ -190,7 +190,7 @@ DWORD ZFileRMDir()
 
 DWORD ZFileGetDir()
 {
-  return(getcwd(DirName,128));
+  return (DWORD) (getcwd(DirName,128));
 }
 
 BYTE * ZFileFindPATH;
